- dbwiddis confirmed from the source code this will be a uint64 value

+++ oshi-core/src/main/java/oshi/hardware/platform/unix/freebsd/FreeBsdDisks.java
@@ -67,7 +67,7 @@ public class FreeBsdDisks implements Disks {
             diskStore.setReadBytes((long) (ParseUtil.parseDoubleOrDefault(split[3], 0d) * 1024));
             diskStore.setWriteBytes((long) (ParseUtil.parseDoubleOrDefault(split[4], 0d) * 1024));
             // # transactions
-            diskStore.setCurrentQueueLength((long) (ParseUtil.parseDoubleOrDefault(split[5], 0d)));
+            diskStore.setCurrentQueueLength(ParseUtil.parseLongOrDefault(split[5], 0L));
             // In seconds, multiply for ms
             diskStore.setTransferTime((long) (ParseUtil.parseDoubleOrDefault(split[6], 0d) * 1000));
             diskStore.setTimeStamp(timeStamp);
@@ -109,7 +109,7 @@ public class FreeBsdDisks implements Disks {
             store.setReadBytes((long) (ParseUtil.parseDoubleOrDefault(split[3], 0d) * 1024));
             store.setWriteBytes((long) (ParseUtil.parseDoubleOrDefault(split[4], 0d) * 1024));
             // # transactions
-            store.setCurrentQueueLength((long) (ParseUtil.parseDoubleOrDefault(split[5], 0d)));
+            store.setCurrentQueueLength(ParseUtil.parseLongOrDefault(split[5], 0L));
             // In seconds, multiply for ms
             store.setTransferTime((long) (ParseUtil.parseDoubleOrDefault(split[6], 0d) * 1000));
             store.setTimeStamp(timeStamp);
